clean: split SQLiteLoader to SQLiteReader and SQLiteWriter to improve readability
