+++ lib/mosaic-database/src/main/java/org/eclipse/mosaic/lib/database/Database.java
@@ -20,8 +20,9 @@ import org.eclipse.mosaic.lib.database.building.Corner;
 import org.eclipse.mosaic.lib.database.building.Wall;
 import org.eclipse.mosaic.lib.database.persistence.DatabaseLoader;
 import org.eclipse.mosaic.lib.database.persistence.OutdatedDatabaseException;
-import org.eclipse.mosaic.lib.database.persistence.SQLiteLoader;
+import org.eclipse.mosaic.lib.database.persistence.SQLiteReader;
 import org.eclipse.mosaic.lib.database.persistence.SQLiteTypeDetector;
+import org.eclipse.mosaic.lib.database.persistence.SQLiteWriter;
 import org.eclipse.mosaic.lib.database.road.Connection;
 import org.eclipse.mosaic.lib.database.road.Node;
 import org.eclipse.mosaic.lib.database.road.Restriction;
@@ -183,7 +184,7 @@ public class Database {
      * @param filename database filename.
      */
     public void saveToFile(String filename) {
-        new SQLiteLoader().saveToFile(this, filename);
+        new SQLiteWriter().saveToFile(this, filename);
     }
 
     /**
@@ -337,16 +338,16 @@ public class Database {
     /**
      * Returns all roundabouts from the database.
      *
-     * @return List of all roundabouts.
+     * @return All roundabouts in a {@link Collection}.
      */
     public Collection<Roundabout> getRoundabouts() {
-        return Collections.unmodifiableList(roundabouts);
+        return Collections.unmodifiableCollection(roundabouts);
     }
 
     /**
      * Returns all buildings from the database.
      *
-     * @return List of all buildings.
+     * @return All buildings in a {@link Collection}.
      */
     public Collection<Building> getBuildings() {
         return Collections.unmodifiableCollection(buildings.values());
@@ -449,7 +450,7 @@ public class Database {
             try {
                 String contentType = Files.probeContentType(file.toPath());
                 if (SQLiteTypeDetector.MIME_TYPE.equals(contentType)) {
-                    loader = new SQLiteLoader();
+                    loader = new SQLiteReader();
                     log.debug("recognized database format is SQLite");
                 } else {
                     loader = null;
@@ -777,12 +778,12 @@ public class Database {
         /**
          * Adds a new building by a given list of corners defining the corners of the building.
          *
-         * @param id the id of the building
-         * @param name the name of the building
-         * @param height the height of the building
-         * @param corners the list geographical points describing the contour of the building
+         * @param id      the id of the building
+         * @param name    the name of the building
+         * @param height  the height of the building
+         * @param corners the list of geographical points describing the contour of the building
          */
-        public Building addBuilding(@Nonnull  String id, String name, double height, GeoPoint[] corners) {
+        public Building addBuilding(@Nonnull String id, String name, double height, GeoPoint[] corners) {
             Validate.isTrue(corners.length > 2, "Building with id " + id + " needs at least three corner points");
 
             final List<Wall> walls = new ArrayList<>();
