+++ lib/mosaic-database/src/main/java/org/eclipse/mosaic/lib/database/persistence/SQLiteAccess.java
@@ -36,9 +36,9 @@ import java.util.function.Function;
  * Static call of SQL statements.
  */
 class SQLiteAccess {
-    protected Connection dbConnection;
-    protected Logger log = LoggerFactory.getLogger(this.getClass());
-    protected String dbName = "";
+    private Connection dbConnection;
+    private Logger log = LoggerFactory.getLogger(this.getClass());
+    private String dbName = "";
 
     /**
      * Constructor for accessing an existing database file.
@@ -52,6 +52,10 @@ class SQLiteAccess {
         } catch (SQLException e) {
             throw new RuntimeException("unexpected error while accessing db file", e);
         }
+    }
+
+    Connection getConnection() {
+        return dbConnection;
     }
 
     /**
